+++ tests/testthat/test-backend-redshift.R
@@ -1,0 +1,4 @@
+test_that("defaults to postgres translations", {
+  local_con(simulate_redshift())
+  expect_equal(translate_sql(log10(x)), sql("LOG(`x`)"))
+})